requirements.txt
runpod
websocket-client
requests
boto3
pyyaml
